Clean up - preping for marquee
As I look to making the `marquee` functionality work, I decided
that the processors should be differently initialized.

This PR updates `ImageProcessor` to take `words` as an initialization
parameter and updates code using the processor.

We also introduce a simple CLI script to test different processing
options.  This first rev takes no args (except words) and generates
a gif.